Removes docs
